+++ src/lv_objx/lv_lmeter.h
@@ -37,7 +37,7 @@ typedef struct
     /*New data for this type */
     uint16_t scale_angle; /*Angle of the scale in deg. (0..360)*/
     uint16_t angle_ofs;
-    uint8_t line_cnt;     /*Count of lines */
+    uint16_t line_cnt;     /*Count of lines */
     int16_t cur_value;
     int16_t min_value;
     int16_t max_value;
@@ -87,7 +87,7 @@ void lv_lmeter_set_range(lv_obj_t * lmeter, int16_t min, int16_t max);
  * @param angle angle of the scale (0..360)
  * @param line_cnt number of lines
  */
-void lv_lmeter_set_scale(lv_obj_t * lmeter, uint16_t angle, uint8_t line_cnt);
+void lv_lmeter_set_scale(lv_obj_t * lmeter, uint16_t angle, uint16_t line_cnt);
 
 /**
  * Set the set an offset for the line meter's angles to rotate it.
@@ -138,7 +138,7 @@ int16_t lv_lmeter_get_max_value(const lv_obj_t * lmeter);
  * @param lmeter pointer to a line meter object
  * @return number of the scale units
  */
-uint8_t lv_lmeter_get_line_count(const lv_obj_t * lmeter);
+uint16_t lv_lmeter_get_line_count(const lv_obj_t * lmeter);
 
 /**
  * Get the scale angle of a line meter
